Detach notes, overrides and severitylevel, also fix timeunit
