+++ gvm/protocols/gmpv208/entities/notes.py
@@ -1,0 +1,269 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2021 Greenbone Networks GmbH
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+# pylint:  disable=redefined-builtin
+# MAYBE we should change filter to filter_string (everywhere)
+
+from typing import Any, List, Optional
+
+from gvm.errors import RequiredArgument, InvalidArgumentType
+from gvm.protocols.gmpv208.entities.severity import Severity, SeverityLevel
+from gvm.utils import add_filter, to_comma_list, to_bool
+from gvm.xml import XmlCommand
+
+
+class NoteMixin:
+    def create_note(
+        self,
+        text: str,
+        nvt_oid: str,
+        *,
+        days_active: Optional[int] = None,
+        hosts: Optional[List[str]] = None,
+        port: Optional[int] = None,
+        result_id: Optional[str] = None,
+        severity: Optional[Severity] = None,
+        task_id: Optional[str] = None,
+        threat: Optional[SeverityLevel] = None,
+    ) -> Any:
+        """Create a new note
+
+        Arguments:
+            text: Text of the new note
+            nvt_id: OID of the nvt to which note applies
+            days_active: Days note will be active. -1 on
+                always, 0 off
+            hosts: A list of hosts addresses
+            port: Port to which the note applies
+            result_id: UUID of a result to which note applies
+            severity: Severity to which note applies
+            task_id: UUID of task to which note applies
+            threat: Severity level to which note applies. Will be converted to
+                severity.
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not text:
+            raise RequiredArgument(
+                function=self.create_note.__name__, argument='text'
+            )
+
+        if not nvt_oid:
+            raise RequiredArgument(
+                function=self.create_note.__name__, argument='nvt_oid'
+            )
+
+        cmd = XmlCommand("create_note")
+        cmd.add_element("text", text)
+        cmd.add_element("nvt", attrs={"oid": nvt_oid})
+
+        if days_active is not None:
+            cmd.add_element("active", str(days_active))
+
+        if hosts:
+            cmd.add_element("hosts", to_comma_list(hosts))
+
+        if port:
+            cmd.add_element("port", str(port))
+
+        if result_id:
+            cmd.add_element("result", attrs={"id": result_id})
+
+        if severity:
+            cmd.add_element("severity", str(severity))
+
+        if task_id:
+            cmd.add_element("task", attrs={"id": task_id})
+
+        if threat is not None:
+            if not isinstance(threat, SeverityLevel):
+                raise InvalidArgumentType(
+                    function="create_note",
+                    argument="threat",
+                    arg_type=SeverityLevel.__name__,
+                )
+
+            cmd.add_element("threat", threat.value)
+
+        return self._send_xml_command(cmd)
+
+    def clone_note(self, note_id: str) -> Any:
+        """Clone an existing note
+
+        Arguments:
+            note_id: UUID of an existing note to clone from
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not note_id:
+            raise RequiredArgument(
+                function=self.clone_note.__name__, argument='note_id'
+            )
+
+        cmd = XmlCommand("create_note")
+        cmd.add_element("copy", note_id)
+        return self._send_xml_command(cmd)
+
+    def delete_note(
+        self, note_id: str, *, ultimate: Optional[bool] = False
+    ) -> Any:
+        """Deletes an existing note
+
+        Arguments:
+            note_id: UUID of the note to be deleted.
+            ultimate: Whether to remove entirely,or to the trashcan.
+        """
+        if not note_id:
+            raise RequiredArgument(
+                function=self.delete_note.__name__, argument='note_id'
+            )
+
+        cmd = XmlCommand("delete_note")
+        cmd.set_attribute("note_id", note_id)
+        cmd.set_attribute("ultimate", to_bool(ultimate))
+
+        return self._send_xml_command(cmd)
+
+    def get_notes(
+        self,
+        *,
+        filter: Optional[str] = None,
+        filter_id: Optional[str] = None,
+        details: Optional[bool] = None,
+        result: Optional[bool] = None,
+    ) -> Any:
+        """Request a list of notes
+
+        Arguments:
+            filter: Filter term to use for the query
+            filter_id: UUID of an existing filter to use for the query
+            details: Add info about connected results and tasks
+            result: Return the details of possible connected results.
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        cmd = XmlCommand("get_notes")
+
+        add_filter(cmd, filter, filter_id)
+
+        if details is not None:
+            cmd.set_attribute("details", to_bool(details))
+
+        if result is not None:
+            cmd.set_attribute("result", to_bool(result))
+
+        return self._send_xml_command(cmd)
+
+    def get_note(self, note_id: str) -> Any:
+        """Request a single note
+
+        Arguments:
+            note_id: UUID of an existing note
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not note_id:
+            raise RequiredArgument(
+                function=self.get_note.__name__, argument='note_id'
+            )
+
+        cmd = XmlCommand("get_notes")
+        cmd.set_attribute("note_id", note_id)
+
+        # for single entity always request all details
+        cmd.set_attribute("details", "1")
+        return self._send_xml_command(cmd)
+
+    def modify_note(
+        self,
+        note_id: str,
+        text: str,
+        *,
+        days_active: Optional[int] = None,
+        hosts: Optional[List[str]] = None,
+        port: Optional[int] = None,
+        result_id: Optional[str] = None,
+        severity: Optional[Severity] = None,
+        task_id: Optional[str] = None,
+        threat: Optional[SeverityLevel] = None,
+    ) -> Any:
+        """Modifies an existing note.
+
+        Arguments:
+            note_id: UUID of note to modify.
+            text: The text of the note.
+            days_active: Days note will be active. -1 on always, 0 off.
+            hosts: A list of hosts addresses
+            port: Port to which note applies.
+            result_id: Result to which note applies.
+            severity: Severity to which note applies.
+            task_id: Task to which note applies.
+            threat: Threat level to which note applies. Will be converted to
+                severity.
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not note_id:
+            raise RequiredArgument(
+                function=self.modify_note.__name__, argument='note_id'
+            )
+
+        if not text:
+            raise RequiredArgument(
+                function=self.modify_note.__name__, argument='text'
+            )
+
+        cmd = XmlCommand("modify_note")
+        cmd.set_attribute("note_id", note_id)
+        cmd.add_element("text", text)
+
+        if days_active is not None:
+            cmd.add_element("active", str(days_active))
+
+        if hosts:
+            cmd.add_element("hosts", to_comma_list(hosts))
+
+        if port:
+            cmd.add_element("port", str(port))
+
+        if result_id:
+            cmd.add_element("result", attrs={"id": result_id})
+
+        if severity:
+            cmd.add_element("severity", str(severity))
+
+        if task_id:
+            cmd.add_element("task", attrs={"id": task_id})
+
+        if threat is not None:
+
+            if not isinstance(threat, SeverityLevel):
+                raise InvalidArgumentType(
+                    function=self.modify_note.__name__,
+                    argument='threat',
+                    arg_type=SeverityLevel.__name__,
+                )
+
+            cmd.add_element("threat", threat.value)
+
+        return self._send_xml_command(cmd)
