+++ gvm/protocols/gmpv208/gmpv208.py
@@ -27,7 +27,7 @@ Module for communication with gvmd in
 """
 import collections
 import logging
-import numbers
+from numbers import Integral
 
 from typing import Any, List, Optional, Callable, Union, Tuple
 from lxml import etree
@@ -59,8 +59,6 @@ PROTOCOL_VERSION = (20, 8)
 
 
 logger = logging.getLogger(__name__)
-
-Severity = numbers.Real
 
 # put this into the Alert Entity
 def _check_event(
@@ -1807,7 +1805,7 @@ class GmpV208Mixin(GvmProtocol):
 
             if schedule_periods is not None:
                 if (
-                    not isinstance(schedule_periods, numbers.Integral)
+                    not isinstance(schedule_periods, Integral)
                     or schedule_periods < 0
                 ):
                     raise InvalidArgument(
@@ -3034,209 +3032,6 @@ class GmpV208Mixin(GvmProtocol):
 
         return self._send_xml_command(cmd)
 
-    def create_note(
-        self,
-        text: str,
-        nvt_oid: str,
-        *,
-        days_active: Optional[int] = None,
-        hosts: Optional[List[str]] = None,
-        port: Optional[int] = None,
-        result_id: Optional[str] = None,
-        severity: Optional[Severity] = None,
-        task_id: Optional[str] = None,
-        threat: Optional[SeverityLevel] = None,
-    ) -> Any:
-        """Create a new note
-
-        Arguments:
-            text: Text of the new note
-            nvt_id: OID of the nvt to which note applies
-            days_active: Days note will be active. -1 on
-                always, 0 off
-            hosts: A list of hosts addresses
-            port: Port to which the note applies
-            result_id: UUID of a result to which note applies
-            severity: Severity to which note applies
-            task_id: UUID of task to which note applies
-            threat: Severity level to which note applies. Will be converted to
-                severity.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not text:
-            raise RequiredArgument(
-                function=self.create_note.__name__, argument='text'
-            )
-
-        if not nvt_oid:
-            raise RequiredArgument(
-                function=self.create_note.__name__, argument='nvt_oid'
-            )
-
-        cmd = XmlCommand("create_note")
-        cmd.add_element("text", text)
-        cmd.add_element("nvt", attrs={"oid": nvt_oid})
-
-        if days_active is not None:
-            cmd.add_element("active", str(days_active))
-
-        if hosts:
-            cmd.add_element("hosts", to_comma_list(hosts))
-
-        if port:
-            cmd.add_element("port", str(port))
-
-        if result_id:
-            cmd.add_element("result", attrs={"id": result_id})
-
-        if severity:
-            cmd.add_element("severity", str(severity))
-
-        if task_id:
-            cmd.add_element("task", attrs={"id": task_id})
-
-        if threat is not None:
-            if not isinstance(threat, SeverityLevel):
-                raise InvalidArgumentType(
-                    function="create_note",
-                    argument="threat",
-                    arg_type=SeverityLevel.__name__,
-                )
-
-            cmd.add_element("threat", threat.value)
-
-        return self._send_xml_command(cmd)
-
-    def clone_note(self, note_id: str) -> Any:
-        """Clone an existing note
-
-        Arguments:
-            note_id: UUID of an existing note to clone from
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not note_id:
-            raise RequiredArgument(
-                function=self.clone_note.__name__, argument='note_id'
-            )
-
-        cmd = XmlCommand("create_note")
-        cmd.add_element("copy", note_id)
-        return self._send_xml_command(cmd)
-
-    def create_override(
-        self,
-        text: str,
-        nvt_oid: str,
-        *,
-        days_active: Optional[int] = None,
-        hosts: Optional[List[str]] = None,
-        port: Optional[int] = None,
-        result_id: Optional[str] = None,
-        severity: Optional[Severity] = None,
-        new_severity: Optional[Severity] = None,
-        task_id: Optional[str] = None,
-        threat: Optional[SeverityLevel] = None,
-        new_threat: Optional[SeverityLevel] = None,
-    ) -> Any:
-        """Create a new override
-
-        Arguments:
-            text: Text of the new override
-            nvt_id: OID of the nvt to which override applies
-            days_active: Days override will be active. -1 on always, 0 off
-            hosts: A list of host addresses
-            port: Port to which the override applies
-            result_id: UUID of a result to which override applies
-            severity: Severity to which override applies
-            new_severity: New severity for result
-            task_id: UUID of task to which override applies
-            threat: Severity level to which override applies. Will be converted
-                to severity.
-            new_threat: New severity level for results. Will be converted to
-                new_severity.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not text:
-            raise RequiredArgument(
-                function=self.create_override.__name__, argument='text'
-            )
-
-        if not nvt_oid:
-            raise RequiredArgument(
-                function=self.create_override.__name__, argument='nvt_oid'
-            )
-
-        cmd = XmlCommand("create_override")
-        cmd.add_element("text", text)
-        cmd.add_element("nvt", attrs={"oid": nvt_oid})
-
-        if days_active is not None:
-            cmd.add_element("active", str(days_active))
-
-        if hosts:
-            cmd.add_element("hosts", to_comma_list(hosts))
-
-        if port:
-            cmd.add_element("port", str(port))
-
-        if result_id:
-            cmd.add_element("result", attrs={"id": result_id})
-
-        if severity:
-            cmd.add_element("severity", str(severity))
-
-        if new_severity:
-            cmd.add_element("new_severity", str(new_severity))
-
-        if task_id:
-            cmd.add_element("task", attrs={"id": task_id})
-
-        if threat is not None:
-            if not isinstance(threat, SeverityLevel):
-                raise InvalidArgumentType(
-                    function=self.create_override.__name__,
-                    argument="threat",
-                    arg_type=SeverityLevel.__name__,
-                )
-
-            cmd.add_element("threat", threat.value)
-
-        if new_threat is not None:
-            if not isinstance(new_threat, SeverityLevel):
-                raise InvalidArgumentType(
-                    function=self.create_override.__name__,
-                    argument="new_threat",
-                    arg_type=SeverityLevel.__name__,
-                )
-
-            cmd.add_element("new_threat", new_threat.value)
-
-        return self._send_xml_command(cmd)
-
-    def clone_override(self, override_id: str) -> Any:
-        """Clone an existing override
-
-        Arguments:
-            override_id: UUID of an existing override to clone from
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not override_id:
-            raise RequiredArgument(
-                function=self.clone_override.__name__, argument='override_id'
-            )
-
-        cmd = XmlCommand("create_override")
-        cmd.add_element("copy", override_id)
-        return self._send_xml_command(cmd)
-
     def clone_permission(self, permission_id: str) -> Any:
         """Clone an existing permission
 
@@ -3686,46 +3481,6 @@ class GmpV208Mixin(GvmProtocol):
 
         cmd = XmlCommand("delete_group")
         cmd.set_attribute("group_id", group_id)
-        cmd.set_attribute("ultimate", to_bool(ultimate))
-
-        return self._send_xml_command(cmd)
-
-    def delete_note(
-        self, note_id: str, *, ultimate: Optional[bool] = False
-    ) -> Any:
-        """Deletes an existing note
-
-        Arguments:
-            note_id: UUID of the note to be deleted.
-            ultimate: Whether to remove entirely,or to the trashcan.
-        """
-        if not note_id:
-            raise RequiredArgument(
-                function=self.delete_note.__name__, argument='note_id'
-            )
-
-        cmd = XmlCommand("delete_note")
-        cmd.set_attribute("note_id", note_id)
-        cmd.set_attribute("ultimate", to_bool(ultimate))
-
-        return self._send_xml_command(cmd)
-
-    def delete_override(
-        self, override_id: str, *, ultimate: Optional[bool] = False
-    ) -> Any:
-        """Deletes an existing override
-
-        Arguments:
-            override_id: UUID of the override to be deleted.
-            ultimate: Whether to remove entirely, or to the trashcan.
-        """
-        if not override_id:
-            raise RequiredArgument(
-                function=self.delete_override.__name__, argument='override_id'
-            )
-
-        cmd = XmlCommand("delete_override")
-        cmd.set_attribute("override_id", override_id)
         cmd.set_attribute("ultimate", to_bool(ultimate))
 
         return self._send_xml_command(cmd)
@@ -4230,111 +3985,6 @@ class GmpV208Mixin(GvmProtocol):
         cmd.set_attribute("group_id", group_id)
         return self._send_xml_command(cmd)
 
-    def get_notes(
-        self,
-        *,
-        filter: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        details: Optional[bool] = None,
-        result: Optional[bool] = None,
-    ) -> Any:
-        """Request a list of notes
-
-        Arguments:
-            filter: Filter term to use for the query
-            filter_id: UUID of an existing filter to use for the query
-            details: Add info about connected results and tasks
-            result: Return the details of possible connected results.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_notes")
-
-        add_filter(cmd, filter, filter_id)
-
-        if details is not None:
-            cmd.set_attribute("details", to_bool(details))
-
-        if result is not None:
-            cmd.set_attribute("result", to_bool(result))
-
-        return self._send_xml_command(cmd)
-
-    def get_note(self, note_id: str) -> Any:
-        """Request a single note
-
-        Arguments:
-            note_id: UUID of an existing note
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not note_id:
-            raise RequiredArgument(
-                function=self.get_note.__name__, argument='note_id'
-            )
-
-        cmd = XmlCommand("get_notes")
-        cmd.set_attribute("note_id", note_id)
-
-        # for single entity always request all details
-        cmd.set_attribute("details", "1")
-        return self._send_xml_command(cmd)
-
-    def get_overrides(
-        self,
-        *,
-        filter: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        details: Optional[bool] = None,
-        result: Optional[bool] = None,
-    ) -> Any:
-        """Request a list of overrides
-
-        Arguments:
-            filter: Filter term to use for the query
-            filter_id: UUID of an existing filter to use for the query
-            details: Whether to include full details
-            result: Whether to include results using the override
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_overrides")
-
-        add_filter(cmd, filter, filter_id)
-
-        if details is not None:
-            cmd.set_attribute("details", to_bool(details))
-
-        if result is not None:
-            cmd.set_attribute("result", to_bool(result))
-
-        return self._send_xml_command(cmd)
-
-    def get_override(self, override_id: str) -> Any:
-        """Request a single override
-
-        Arguments:
-            override_id: UUID of an existing override
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_overrides")
-
-        if not override_id:
-            raise RequiredArgument(
-                function=self.get_override.__name__, argument='override_id'
-            )
-
-        cmd.set_attribute("override_id", override_id)
-
-        # for single entity always request all details
-        cmd.set_attribute("details", "1")
-        return self._send_xml_command(cmd)
-
     def get_permissions(
         self,
         *,
@@ -4733,7 +4383,7 @@ class GmpV208Mixin(GvmProtocol):
             cmd.set_attribute("name", name)
 
         if duration is not None:
-            if not isinstance(duration, numbers.Integral):
+            if not isinstance(duration, Integral):
                 raise InvalidArgument("duration needs to be an integer number")
 
             cmd.set_attribute("duration", str(duration))
@@ -5280,172 +4930,6 @@ class GmpV208Mixin(GvmProtocol):
 
         if users:
             cmd.add_element("users", to_comma_list(users))
-
-        return self._send_xml_command(cmd)
-
-    def modify_note(
-        self,
-        note_id: str,
-        text: str,
-        *,
-        days_active: Optional[int] = None,
-        hosts: Optional[List[str]] = None,
-        port: Optional[int] = None,
-        result_id: Optional[str] = None,
-        severity: Optional[Severity] = None,
-        task_id: Optional[str] = None,
-        threat: Optional[SeverityLevel] = None,
-    ) -> Any:
-        """Modifies an existing note.
-
-        Arguments:
-            note_id: UUID of note to modify.
-            text: The text of the note.
-            days_active: Days note will be active. -1 on always, 0 off.
-            hosts: A list of hosts addresses
-            port: Port to which note applies.
-            result_id: Result to which note applies.
-            severity: Severity to which note applies.
-            task_id: Task to which note applies.
-            threat: Threat level to which note applies. Will be converted to
-                severity.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not note_id:
-            raise RequiredArgument(
-                function=self.modify_note.__name__, argument='note_id'
-            )
-
-        if not text:
-            raise RequiredArgument(
-                function=self.modify_note.__name__, argument='text'
-            )
-
-        cmd = XmlCommand("modify_note")
-        cmd.set_attribute("note_id", note_id)
-        cmd.add_element("text", text)
-
-        if days_active is not None:
-            cmd.add_element("active", str(days_active))
-
-        if hosts:
-            cmd.add_element("hosts", to_comma_list(hosts))
-
-        if port:
-            cmd.add_element("port", str(port))
-
-        if result_id:
-            cmd.add_element("result", attrs={"id": result_id})
-
-        if severity:
-            cmd.add_element("severity", str(severity))
-
-        if task_id:
-            cmd.add_element("task", attrs={"id": task_id})
-
-        if threat is not None:
-
-            if not isinstance(threat, SeverityLevel):
-                raise InvalidArgumentType(
-                    function=self.modify_note.__name__,
-                    argument='threat',
-                    arg_type=SeverityLevel.__name__,
-                )
-
-            cmd.add_element("threat", threat.value)
-
-        return self._send_xml_command(cmd)
-
-    def modify_override(
-        self,
-        override_id: str,
-        text: str,
-        *,
-        days_active: Optional[int] = None,
-        hosts: Optional[List[str]] = None,
-        port: Optional[int] = None,
-        result_id: Optional[str] = None,
-        severity: Optional[Severity] = None,
-        new_severity: Optional[Severity] = None,
-        task_id: Optional[str] = None,
-        threat: Optional[SeverityLevel] = None,
-        new_threat: Optional[SeverityLevel] = None,
-    ) -> Any:
-        """Modifies an existing override.
-
-        Arguments:
-            override_id: UUID of override to modify.
-            text: The text of the override.
-            days_active: Days override will be active. -1 on always,
-                0 off.
-            hosts: A list of host addresses
-            port: Port to which override applies.
-            result_id: Result to which override applies.
-            severity: Severity to which override applies.
-            new_severity: New severity score for result.
-            task_id: Task to which override applies.
-            threat: Threat level to which override applies.
-                Will be converted to severity.
-            new_threat: New threat level for results. Will be converted to
-                new_severity.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not override_id:
-            raise RequiredArgument(
-                function=self.modify_override.__name__, argument='override_id'
-            )
-        if not text:
-            raise RequiredArgument(
-                function=self.modify_override.__name__, argument='text'
-            )
-
-        cmd = XmlCommand("modify_override")
-        cmd.set_attribute("override_id", override_id)
-        cmd.add_element("text", text)
-
-        if days_active is not None:
-            cmd.add_element("active", str(days_active))
-
-        if hosts:
-            cmd.add_element("hosts", to_comma_list(hosts))
-
-        if port:
-            cmd.add_element("port", str(port))
-
-        if result_id:
-            cmd.add_element("result", attrs={"id": result_id})
-
-        if severity:
-            cmd.add_element("severity", str(severity))
-
-        if new_severity:
-            cmd.add_element("new_severity", str(new_severity))
-
-        if task_id:
-            cmd.add_element("task", attrs={"id": task_id})
-
-        if threat is not None:
-            if not isinstance(threat, SeverityLevel):
-                raise InvalidArgumentType(
-                    function=self.modify_override.__name__,
-                    argument='threat',
-                    arg_type=SeverityLevel.__name__,
-                )
-            cmd.add_element("threat", threat.value)
-
-        if new_threat is not None:
-            if not isinstance(new_threat, SeverityLevel):
-                raise InvalidArgumentType(
-                    function=self.modify_override.__name__,
-                    argument='new_threat',
-                    arg_type=SeverityLevel.__name__,
-                )
-
-            cmd.add_element("new_threat", new_threat.value)
 
         return self._send_xml_command(cmd)
